Refactor & simplify job execution
