+++ backend/src/forecastbox/api/execution.py
@@ -26,7 +26,7 @@ from fastapi import HTTPException
 from pydantic import BaseModel
 
 from forecastbox.api.artifacts.manager import ArtifactManager, submit_artifact_download
-from forecastbox.api.types import EnvironmentSpecification, ExecutionSpecification, ForecastProducts, RawCascadeJob
+from forecastbox.api.types.jobs import EnvironmentSpecification, ExecutionSpecification, RawCascadeJob
 from forecastbox.api.utils import get_model_path
 from forecastbox.config import config
 from forecastbox.db.job import insert_one
@@ -37,73 +37,15 @@ from forecastbox.schemas.user import UserRead
 logger = logging.getLogger(__name__)
 
 
-def _get_model(spec: ForecastProducts):
-    model_spec = dict(
-        lead_time=spec.model.lead_time,
-        date=spec.model.date,
-        ensemble_members=spec.model.ensemble_members,
-    )
-
-    return get_model(checkpoint=get_model_path(spec.model.model)).specify(**model_spec)  # type: ignore
-
-
+# TODO replace with just output_ids, or something more fitting, since product_ are deprecated
 class ProductToOutputId(BaseModel):
     product_name: str
     product_spec: dict[str, Any]
     output_ids: Sequence[str]
 
 
-def forecast_products_to_cascade(
-    spec: ForecastProducts, environment: EnvironmentSpecification
-) -> tuple[Cascade, dict, list[ProductToOutputId]]:
-    model = _get_model(spec)
-
-    # Create the model action graph
-    model_action = model.graph()
-
-    # Iterate over each product in the specification
-    complete_graph = Graph([])
-
-    product_to_id_mappings = []
-
-    for product in spec.products:
-        product_spec = product.specification.copy()
-        try:
-            product_graph = get_product(*product.product.split("/", 1)).execute(product_spec, model, model_action)
-        except Exception as e:
-            raise Exception(f"Error in product {product}:\n{e}")
-
-        if isinstance(product_graph, fluent.Action):
-            product_graph = product_graph.graph()
-
-        output_ids = [x.name for x in product_graph.sinks]
-        product_to_id_mappings.append(ProductToOutputId(product_name=product.product, product_spec=product_spec, output_ids=output_ids))
-
-        complete_graph += product_graph
-
-    if len(spec.products) == 0:
-        complete_graph += model_action.graph()
-        product_to_id_mappings.append(
-            ProductToOutputId(product_name="All Model Outputs", product_spec={}, output_ids=[x.name for x in model_action.graph().sinks])
-        )
-
-    env_vars = model.control.environment_variables
-    env_vars.update(environment.environment_variables)
-
-    return Cascade(deduplicate_nodes(complete_graph)), env_vars, product_to_id_mappings
-
-
-def execution_specification_to_cascade(spec: ExecutionSpecification) -> tuple[JobInstance, dict, list[ProductToOutputId] | None]:
-    if isinstance(spec.job, ForecastProducts):
-        cascade_graph, environment_variables, product_to_id_mappings = forecast_products_to_cascade(spec.job, spec.environment)
-        return graph2job(cascade_graph._graph), environment_variables, product_to_id_mappings
-    elif isinstance(spec.job, RawCascadeJob):
-        return spec.job.job_instance, {}, None
-    assert_never(spec.job)
-
-
 def _execute_cascade(spec: ExecutionSpecification) -> tuple[api.SubmitJobResponse, list[ProductToOutputId]]:
-    """Converts spec to JobInstance and submits to cascade api, returning response + list of sinks"""
+    """Converts spec to JobInstance and submits to cascade api, returning response"""
 
     # Handle runtime artifacts download before job execution
     runtime_artifacts = spec.environment.runtime_artifacts
@@ -144,18 +86,11 @@ def _execute_cascade(spec: ExecutionSpecification) -> tuple[api.SubmitJobRespons
 
                 time.sleep(1)
 
-    try:
-        job, job_envvars, product_to_id_mappings = execution_specification_to_cascade(spec)
-    except Exception as e:
-        return api.SubmitJobResponse(job_id=None, error=repr(e)), []
-
+    job = spec.job.job_instance
     sinks = cascade_views.sinks(job)
     sinks = [s for s in sinks if not s.task.startswith("run_as_earthkit")]
-
-    if not product_to_id_mappings:
-        product_to_id_mappings = [ProductToOutputId(product_name="All Outputs", product_spec={}, output_ids=[x.task for x in sinks])]
-
     job.ext_outputs = sinks
+    product_to_id_mappings = [ProductToOutputId(product_name="All Outputs", product_spec={}, output_ids=[x.task for x in sinks])]
 
     environment = spec.environment
 
@@ -163,7 +98,6 @@ def _execute_cascade(spec: ExecutionSpecification) -> tuple[api.SubmitJobRespons
     workers_per_host = min(config.cascade.max_workers_per_host, environment.workers_per_host or config.cascade.max_workers_per_host)
 
     env_vars = {"TMPDIR": config.cascade.venv_temp_dir}
-    env_vars.update({k: str(v) for k, v in job_envvars.items()})
 
     r = api.SubmitJobRequest(
         job=api.JobSpec(
