+++ backend/src/forecastbox/api/routers/fable.py
@@ -20,8 +20,8 @@ from fiab_core.fable import BlockFactoryCatalogue
 import forecastbox.api.fable as api_fable
 import forecastbox.db.fable as db_fable
 from forecastbox.api.plugin.manager import PluginCompositeId, catalogue_view, plugins_ready
-from forecastbox.api.types import RawCascadeJob
 from forecastbox.api.types.fable import FableBuilderV1, FableValidationExpansion
+from forecastbox.api.types.jobs import RawCascadeJob
 from forecastbox.auth.users import current_active_user
 from forecastbox.schemas.user import UserRead
