+++ backend/src/forecastbox/api/scheduling/job_utils.py
@@ -17,7 +17,7 @@ import orjson
 from cascade.low.func import Either
 
 from forecastbox.api.scheduling.dt_utils import calculate_next_run
-from forecastbox.api.types import ExecutionSpecification
+from forecastbox.api.types.jobs import ExecutionSpecification
 from forecastbox.db.schedule import get_schedules, max_attempt_cnt, run2date, run2schedule
