--- backend/src/forecastbox/api/types/base.py
@@ -1,5 +1,0 @@
-from pydantic import BaseModel
-
-
-class FIABBaseModel(BaseModel):
-    pass
